+++ src/main/java/com/proyectoIntegrador/consultorioOdontologico/service/UsuarioService.java
@@ -2,6 +2,7 @@ package com.proyectoIntegrador.consultorioOdontologico.service;
 
 import com.proyectoIntegrador.consultorioOdontologico.entity.Usuario;
 import com.proyectoIntegrador.consultorioOdontologico.repository.UsuarioRepository;
+import com.proyectoIntegrador.consultorioOdontologico.util.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -26,7 +27,7 @@ public class UsuarioService {
     }
 
     public void eliminar(Integer id) {
-        //if (!usuarioRepository.existsById(id)) throw new ResourceNotFoundException(id.toString(), "Usuario Id");
+        if (!usuarioRepository.existsById(id)) throw new ResourceNotFoundException(id.toString(), "Usuario Id");
 
         usuarioRepository.deleteById(id);
     }
